refactor(1013): extract EnvironmentUtils to src/utils/ (SC-001, position 33)
Fresh extraction of the 119-line EnvironmentUtils class to
src/utils/environment_utils.py using the SC-001 facade pattern.

Notable: this class is self-contained (only stdlib os/logging + inline pwd).
No live globals are touched, so the lazy mh = importlib.import_module
accessor typical of Cat B extractions is not needed.

- Re-export alias added in MistHelper.py alphabetically among src.utils.*
- Class body replaced with single-line breadcrumb comment
- Coverage omit entry added; wave-2 mypy override not required
- Local gate CLEAN: black + ruff + py_compile + --test (66/0/131 baseline)
