+++ pyproject.toml
@@ -253,6 +253,7 @@ omit = [
     "src/site/bulk_radius_wlan_config_manager.py",
     "src/ssh/cli_shell_manager.py",
     "src/ui/tui.py",
+    "src/utils/environment_utils.py",
     "src/websocket/*"
 ]
